refactor: sw_bls12377 hints
